Update README.md
# Breed_recommendation
A user gets recommendations on a dog breed based on their age and lifestyle via a telegram bot.
The user answers three questions: name, age, lifestyle. The last question provides the user with two options: active and passive. The user's answers are converted into points.
There are three categories of dogs denoting their size: big, medium and small. The category that gets recommended to the user depends on the number of the points they have scored. The user who is between 35 and 60 y.o. and has an active lifestyle can achieve the maximum number of points.
Having answered the questions the user finds out the category of the dogs (+some breeds) that would be appropriate for them. Then the user is suggested being showed one breed that is chosen randomly from the given category list. The user can ask the bot to choose another breed.